+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 
 # 0.0.1sink
+- add `list-connectors` `list-config` commands (in addition to `help`, `version`)
 - add preliminary support for AWS SQS
 - add `--config` option
 - add `--memprof` `--cpuprof` options
 - verify configuration file connectors attribute, fail to start if not found
 - remove nasral package (need to find an alternative for env variable)
 - add a new log module to support traceability (incomplete)
--